refactor: 🧩 Move Delete fetcher file and move methods to utils and collections
